+++ CHANGELOG.md
@@ -1,14 +1,27 @@
 
 # FlexJSON Change Log
 
-## 1.2.0
+## 0.3.3 (2026-05-29 unreleased)
+- performance fixes
 
+## 0.3.2 (2026-05-29 unreleased)
+- performance fixes
 
+## 0.3.1 (2026-05-29 unreleased)
+- performance fixes
 
-## 1.1.0 (2026-05-29 unreleased)
+## 0.3.0 (2026-05-29 unreleased)
+
+- iterative parser
+
+## 0.2.0 (2026-05-29 unreleased)
+
+- recursive parser
+
+## 0.1.1 (2026-05-29 unreleased)
 
 - MVP complete
 
-## 1.0.0 (2026-05-28 unreleased)
+## 0.1.0 (2026-05-28 unreleased)
 
-- Initial release
+- Initial version